cleanup and reorganize files for projects and experience
