+++ src/App.jsx
@@ -1,7 +1,7 @@
 import "./App.css";
 import Hero from "./components/layout/hero/Hero";
 import About from "./components/layout/about/About";
-import ExperienceSection from "./components/layout/experience/Experience";
+import Experience from "./components/layout/experience/Experience";
 import Projects from "./components/layout/projects/Projects";
 import Footer from "./components/layout/footer/Footer";
 
@@ -11,7 +11,7 @@ function App() {
       <main>
         <Hero />
         <About />
-        <ExperienceSection />
+        <Experience />
         <Projects />
         <Footer />
       </main>
