--- src/components/layout/projects/Professional.jsx
@@ -1,14 +1,0 @@
-import ProjectTemplate from "./ProjectTemplate";
-import { professionalProjects } from "./data/projects";
-
-function ProfessionalProjects() {
-  return (
-    <>
-      <h3>Professional</h3>
-
-      <ProjectTemplate projects={professionalProjects} />
-    </>
-  );
-}
-
-export default ProfessionalProjects;
